[MetaService] Calling multiple service from the Meta
