+++ Services/gRPC/CNN_ImageRecon/buildproto.sh
