+++ Services/gRPC/CNN_ImageRecon/config/snetd_image_recon_service.json
