+++ Services/gRPC/CNN_ImageRecon/service/common.py
