+++ Services/gRPC/CNN_ImageRecon/service/image_recon_service.py
