+++ Services/gRPC/CNN_ImageRecon/service/model/image_recon.proto
