+++ Services/gRPC/CNN_ImageRecon/service/model/image_recon_pb2_grpc.py
