+++ Services/gRPC/CNN_ImageRecon/test_image_recon_service.py
